Variety of updates
Code cleanup
Make PKs Longs
DAO updates
Rename FAB to ActionButton

+++ composeApp/src/commonMain/kotlin/com/steeplesoft/giftbook/database/dao/GiftIdeaDao.kt
@@ -24,8 +24,8 @@ interface GiftIdeaDao {
         WHERE g.recipientId = :recipId
           AND (g.occasionId IS NULL OR g.occasionId = :occasionId)
     """)
-    suspend fun getCurrentGiftIdeasForRecipAndOccasion(recipId: Int, occasionId: Int): List<GiftIdea>
+    suspend fun getCurrentGiftIdeasForRecipAndOccasion(recipId: Long, occasionId: Long): List<GiftIdea>
 
     @Query("SELECT g.* FROM giftidea g WHERE g.recipientId = :recipId AND g.occasionId IS NOT NULL")
-    suspend fun getUsedGiftIdeasForRecip(recipId: Int): List<GiftIdea>
+    suspend fun getUsedGiftIdeasForRecip(recipId: Long): List<GiftIdea>
 }
